+++ src/app/__root.jsx
@@ -1,0 +1,17 @@
+import { createRootRoute, Outlet } from '@tanstack/react-router';
+import { TanStackRouterDevtools } from '@tanstack/react-router-devtools';
+import '../css/App.css';
+import Header from '../components/Header';
+
+export const Route = createRootRoute({
+  component: () => (
+    <>
+      <main className='app'>
+        <Header />
+      </main>
+      <hr />
+      <Outlet />
+      <TanStackRouterDevtools />
+    </>
+  ),
+});